Added Search and Make Playlist
I wanted the ability to load a playlist from a csv file. This enabled me to easily transcribe the songs in a spreadsheet, check for data entry errors, and then export. I put a time delay so that I don't trigger any rate limiting on the API. May not be required.